chore: wrap NavbarItem instead of replace it
